+++ packages/platform-android/package.json
@@ -13,6 +13,22 @@
     ".": {
       "types": "./src/index.ts",
       "default": "./src/index.ts"
+    },
+    "./adb-executor": {
+      "types": "./src/adb-executor.ts",
+      "default": "./src/adb-executor.ts"
+    },
+    "./adb-host": {
+      "types": "./src/adb-host.ts",
+      "default": "./src/adb-host.ts"
+    },
+    "./ime-helper": {
+      "types": "./src/ime-helper.ts",
+      "default": "./src/ime-helper.ts"
+    },
+    "./ime-lifecycle": {
+      "types": "./src/ime-lifecycle.ts",
+      "default": "./src/ime-lifecycle.ts"
     }
   }
 }
