+++ scripts/layering/package-boundaries.test.ts
@@ -50,9 +50,11 @@ const CONTRACT_EXPORTS = [
   '@agent-device/contracts/alert-contract',
   '@agent-device/contracts/alert-runtime',
   '@agent-device/contracts/android-clipboard-support',
+  '@agent-device/contracts/android-helper-artifacts',
   '@agent-device/contracts/android-input-ownership',
   '@agent-device/contracts/android-snapshot-quality',
   '@agent-device/contracts/android-system-chrome',
+  '@agent-device/contracts/android-touch-plan',
   '@agent-device/contracts/app-deployment-runtime',
   '@agent-device/contracts/app-deployment-runtime-plan',
   '@agent-device/contracts/app-event-runtime',
