+++ src/platforms/android/instrumentation-helper.ts
@@ -1,4 +1,3 @@
-import { AppError } from '@agent-device/kernel/errors';
 
 // Shared primitives for the Android instrumentation helpers (snapshot + multi-touch).
 // Both helpers drive `am instrument -w` and parse the resulting
@@ -104,53 +103,9 @@ export function readInstrumentationResultBoolean(value: string | undefined): boo
   return undefined;
 }
 
-export function readAndroidHelperManifestInteger(
-  value: unknown,
-  field: string,
-  helperLabel: string,
-): number {
-  if (typeof value !== 'number' || !Number.isInteger(value)) {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Android ${helperLabel} manifest ${field} must be an integer.`,
-    );
-  }
-  return value;
-}
-
-export function readAndroidHelperManifestLiteral<const Value extends string>(
-  value: unknown,
-  field: string,
-  expected: Value,
-  helperLabel: string,
-): Value {
-  if (value !== expected) {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Android ${helperLabel} manifest ${field} must be "${expected}".`,
-    );
-  }
-  return expected;
-}
-
-export function readAndroidHelperManifestString(
-  value: unknown,
-  field: string,
-  helperLabel: string,
-): string {
-  if (typeof value !== 'string' || value.length === 0) {
-    throw new AppError('INVALID_ARGS', `Android ${helperLabel} manifest ${field} is required.`);
-  }
-  return value;
-}
-
-export function readAndroidHelperManifestSha256(value: unknown, helperLabel: string): string {
-  const sha256 = readAndroidHelperManifestString(value, 'sha256', helperLabel).trim().toLowerCase();
-  if (sha256.length !== 64 || !/^[0-9a-f]+$/.test(sha256)) {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Android ${helperLabel} manifest sha256 must be a 64-character hex string.`,
-    );
-  }
-  return sha256;
-}
+export {
+  readAndroidHelperManifestInteger,
+  readAndroidHelperManifestLiteral,
+  readAndroidHelperManifestSha256,
+  readAndroidHelperManifestString,
+} from '@agent-device/contracts/android-helper-artifacts';
